pyproject.toml
[tool.poetry]
name = "stockpulse"
version = "0.1.0"
description = ""
authors = ["KevinGregory <kevingregory604@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "< 0.116.0"
solara = "< 1.45.0"
uvicorn = "< 0.35.0"
yfinance = "< 0.3.0"
redis = "< 5.3.1"
fastapi-cache2 = {extras = ["redis"], version = "< 0.3.0"}
pandas = "< 2.3.0"
pydantic = "< 2.11.0"
pydantic-settings = "< 2.8.0"


[tool.poetry.group.dev.dependencies]
ruff = "< 0.10.0"
pytest = "< 8.4.0"
pyright = "< 1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
